+++ include/C/terminal.c
@@ -1,0 +1,70 @@
+#include <VGA.h>
+#include <terminal.h>
+#include <string.h>
+#include <stddef.h>
+
+void terminal_initialize(void) {
+	terminal_row = 0;
+	terminal_column = 0;
+	terminal_color = vga_entry_color(VGA_COLOR_LIGHT_GREY, VGA_COLOR_BLACK);
+	terminal_buffer = (uint16_t*) 0xB8000;
+	for (size_t y = 0; y < VGA_HEIGHT; y++) {
+		for (size_t x = 0; x < VGA_WIDTH; x++) {
+			const size_t index = y * VGA_WIDTH + x;
+			terminal_buffer[index] = vga_entry(' ', terminal_color);
+		}
+	}
+}
+
+void terminal_scroll() {
+  for (size_t i = 0; i < VGA_HEIGHT; i++){
+    for (size_t m = 0; m < VGA_WIDTH; m++){
+        terminal_buffer[i * VGA_WIDTH + m] = terminal_buffer[(i + 1) * VGA_WIDTH + m];
+    }
+  }
+}
+
+void terminal_setcolor(uint8_t color) {
+	terminal_color = color;
+}
+
+void terminal_putentryat(char c, uint8_t color, size_t x, size_t y) {
+	if (c == '\n') {
+		terminal_row++;
+		terminal_column = -1;
+		return;
+	}
+
+	const size_t index = y * VGA_WIDTH + x;
+	terminal_buffer[index] = vga_entry(c, color);
+}
+
+void terminal_putchar(char c) {
+	terminal_putentryat(c, terminal_color, terminal_column, terminal_row);
+	if (++terminal_column == VGA_WIDTH) {
+		terminal_column = 0;
+    terminal_row++;
+	}
+
+	if (terminal_row == VGA_HEIGHT) {
+		terminal_scroll();
+		terminal_row = VGA_HEIGHT - 1;
+	}
+}
+
+void terminal_write(const char* data, size_t size) {
+	for (size_t i = 0; i < size; i++)
+		terminal_putchar(data[i]);
+}
+
+void terminal_writestring(const char* data) {
+	terminal_write(data, strlen(data));
+}
+
+void terminal_clear() {
+	for (size_t x = 0; x < VGA_WIDTH; x++) {
+		for (size_t y = 0; y < VGA_HEIGHT; y++) {
+			terminal_putentryat(' ', terminal_color, x, y);
+		}
+	}
+}
