+++ include/headers/string.h
@@ -1,0 +1,10 @@
+#ifndef STRING_H
+#define STRING_H
+
+size_t strlen(const char* str);
+
+void reverseString(char *ptr, size_t length);
+
+void toString(char *ptr, int length, int t);
+
+#endif
