+++ include/headers/terminal.h
@@ -1,0 +1,28 @@
+#ifndef TERMINAL_H
+#define TERMINAL_H
+
+#include <stdint.h>
+#include <VGA.h>
+
+size_t terminal_row;
+size_t terminal_column;
+uint8_t terminal_color;
+uint16_t* terminal_buffer;
+
+void terminal_initialize(void);
+
+void terminal_scroll();
+
+void terminal_setcolor(uint8_t color);
+
+void terminal_putentryat(char c, uint8_t color, size_t x, size_t y);
+
+void terminal_putchar(char c);
+
+void terminal_write(const char* data, size_t size);
+
+void terminal_writestring(const char* data);
+
+void terminal_clear();
+
+#endif
