+++ kernel.c
@@ -1,7 +1,7 @@
 #include <stdbool.h>
 #include <stddef.h>
 #include <stdint.h>
-#include <memory.h>
+#include <terminal.h>
 
 /* Check if the compiler thinks you are targeting the wrong operating system. */
 #if defined(__linux__)
@@ -14,145 +14,9 @@
 #endif
 
 /* Hardware text mode color constants. */
-enum vga_color {
-	VGA_COLOR_BLACK = 0,
-	VGA_COLOR_BLUE = 1,
-	VGA_COLOR_GREEN = 2,
-	VGA_COLOR_CYAN = 3,
-	VGA_COLOR_RED = 4,
-	VGA_COLOR_MAGENTA = 5,
-	VGA_COLOR_BROWN = 6,
-	VGA_COLOR_LIGHT_GREY = 7,
-	VGA_COLOR_DARK_GREY = 8,
-	VGA_COLOR_LIGHT_BLUE = 9,
-	VGA_COLOR_LIGHT_GREEN = 10,
-	VGA_COLOR_LIGHT_CYAN = 11,
-	VGA_COLOR_LIGHT_RED = 12,
-	VGA_COLOR_LIGHT_MAGENTA = 13,
-	VGA_COLOR_LIGHT_BROWN = 14,
-	VGA_COLOR_WHITE = 15,
-};
-
-static inline uint8_t vga_entry_color(enum vga_color fg, enum vga_color bg) {
-	return fg | bg << 4;
-}
-
-static inline uint16_t vga_entry(unsigned char uc, uint8_t color) {
-	return (uint16_t) uc | (uint16_t) color << 8;
-}
-
-size_t strlen(const char* str) {
-	size_t len = 0;
-	while (str[len])
-		len++;
-	return len;
-}
-
-
-static const size_t VGA_WIDTH = 80;
-static const size_t VGA_HEIGHT = 25;
-
-size_t terminal_row;
-size_t terminal_column;
-uint8_t terminal_color;
-uint16_t* terminal_buffer;
-
-void terminal_initialize(void) {
-	terminal_row = 0;
-	terminal_column = 0;
-	terminal_color = vga_entry_color(VGA_COLOR_LIGHT_GREY, VGA_COLOR_BLACK);
-	terminal_buffer = (uint16_t*) 0xC03FF000;
-	for (size_t y = 0; y < VGA_HEIGHT; y++) {
-		for (size_t x = 0; x < VGA_WIDTH; x++) {
-			const size_t index = y * VGA_WIDTH + x;
-			terminal_buffer[index] = vga_entry(' ', terminal_color);
-		}
-	}
-}
-
-void reverseString(char *ptr, size_t length) {
-  char ca[length];
-
-  for (size_t i = 0; i < length; i++) {
-    ca[(length - 1) - i] = ptr[i];
-  }
-
-  memcpy(ptr, ca, length);
-}
-
-
-void toString(char *ptr, int length, int t) {
-  char ca[length];
-
-  for (int i = 0; i < length; i++) {
-    ca[i] = (char) ((t % 10) + 48);
-    t /= 10;
-    if (t == 0) break;
-  }
-
-  reverseString(ca, strlen(ca));
-
-  memcpy(ptr, ca, length);
-}
-
-void terminal_scroll() {
-  for (size_t i = 0; i < VGA_HEIGHT; i++){
-    for (size_t m = 0; m < VGA_WIDTH; m++){
-        terminal_buffer[i * VGA_WIDTH + m] = terminal_buffer[(i + 1) * VGA_WIDTH + m];
-    }
-  }
-}
-
-void terminal_setcolor(uint8_t color) {
-	terminal_color = color;
-}
-
-void terminal_putentryat(char c, uint8_t color, size_t x, size_t y) {
-	if (c == '\n') {
-		terminal_row++;
-		terminal_column = -1;
-		return;
-	}
-
-	const size_t index = y * VGA_WIDTH + x;
-	terminal_buffer[index] = vga_entry(c, color);
-}
-
-void terminal_putchar(char c) {
-	terminal_putentryat(c, terminal_color, terminal_column, terminal_row);
-	if (++terminal_column == VGA_WIDTH) {
-		terminal_column = 0;
-    terminal_row++;
-	}
-
-	if (terminal_row == VGA_HEIGHT) {
-		terminal_scroll();
-		terminal_row = VGA_HEIGHT - 1;
-	}
-}
-
-void terminal_write(const char* data, size_t size) {
-	for (size_t i = 0; i < size; i++)
-		terminal_putchar(data[i]);
-}
-
-void terminal_writestring(const char* data) {
-	terminal_write(data, strlen(data));
-}
-
-void terminal_clear() {
-	for (size_t x = 0; x < VGA_WIDTH; x++) {
-		for (size_t y = 0; y < VGA_HEIGHT; y++) {
-			terminal_putentryat(' ', terminal_color, x, y);
-		}
-	}
-}
 
 void kernel_main(void) {
 	terminal_initialize();
 
 	terminal_writestring("Initialized Kernel...");
-	int a = 10;
-	int *ptr = (int *)0x01000000;
-	*ptr = 12;
 }
